+++ src/kernel/ram_disk/ram_disk.c
@@ -8,21 +8,21 @@ RamDirectory* rootDir;
 
 void print_directory(RamDirectory* directory, uint64_t indentation)
 {
-    for (int j = 0; j < indentation * 4; j++)
+    for (uint64_t j = 0; j < indentation * 4; j++)
     {
         tty_put(' ');
     }
     tty_print(directory->name);
     tty_put('\n');
 
-    for (int i = 0; i < directory->directoryAmount; i++)
+    for (uint64_t i = 0; i < directory->directoryAmount; i++)
     {
         print_directory(&directory->directories[i], indentation + 1);
     }
 
-    for (int i = 0; i < directory->fileAmount; i++)
+    for (uint64_t i = 0; i < directory->fileAmount; i++)
     {
-        for (int j = 0; j < (indentation + 1) * 4; j++)
+        for (uint64_t j = 0; j < (indentation + 1) * 4; j++)
         {
             tty_put(' ');
         }
@@ -43,7 +43,7 @@ void ram_disk_init(RamDirectory* rootDirectory)
 uint8_t ram_disk_compare_names(const char* nameStart, const char* nameEnd, const char* otherName)
 {
     uint64_t otherNameLength = strlen(otherName);
-    if (otherNameLength != nameEnd - nameStart)
+    if (otherNameLength != (uint64_t)nameEnd - (uint64_t)nameStart)
     {
         return 0;
     }
@@ -80,7 +80,7 @@ RamFile* ram_disk_get(const char* path)
         
             if (path[index] == '/')
             {
-                for (int i = 0; i < currentDir->directoryAmount; i++)
+                for (uint64_t i = 0; i < currentDir->directoryAmount; i++)
                 {
                     if (ram_disk_compare_names(path + prevIndex, path + index, currentDir->directories[i].name))
                     {
@@ -93,7 +93,7 @@ RamFile* ram_disk_get(const char* path)
             }
             else if (path[index] == '\0')
             {                
-                for (int i = 0; i < currentDir->fileAmount; i++)
+                for (uint64_t i = 0; i < currentDir->fileAmount; i++)
                 {
                     if (ram_disk_compare_names(path + prevIndex, path + index, currentDir->files[i].name))
                     {
